Remove unused using statement and ToolTip_OnOpened method
We cleaned up unnecessary using statements from ButtonShelfBook.xaml.cs. The 'System' namespace was not used anywhere in the code, so it’s removal is a part of the cleanup.

Also, we removed the ToolTip_OnOpened method as it was just outputting content to console which isn't a requirement in our production code. This helps to keep the codebase cleaner and easier to maintain.