Add optional book parameter in AddEditBook constructor
The constructor of the AddEditBook class has been modified to accept an optional book parameter. The change enables pre-population of the BookData property when an instance is created with a given book. If the passed book is null, it defaults to a new Book instance. This change aids in reducing boilerplate code, especially when editing a book's details, since we don't have to populate the BookData property manually.